Login (#1)
* Sample Activity

Main Activity displays "Hello Android" on the screen

* Added login activity and associated fragments

there is now the main login activity which handles the 3 different current fragments. These fragments are starting screen, login, and create account

* Update start screen button id

updated button name to not be the same as another

* Added account creation functionality

you can now create an account with a username and password. Currently only checks against hard coded usernames and approves any password

* Updated Password Handling and UI

now description is provided for usernames and passwords and error messages displayed to user if they are not met.

* Password description fix

removed special character requirement

* Added LoginManager

login manager is a singleton which handles checking of usernames and passwords as well as hashing passwords. The goal was to abstract away some of the logic.

* Added login page and demo functionality

* Created title string for login page

* Removed old start screen

the new start screen is the login fragment

* Updated login logic

Now attempting to login uses the login manager.

* Integrated FireBase's FireStore

The app is now able to connect to Firebase's FireStore and properly store and retrieve usernames, hashed passwords, and salt values.

* Added 'Remember Me' functionality

login screen now has a remember me. This works by creating a token with temporary time period of use. Token is stored on firebase and locally.

* Removed hardcoded strings

moved the strings to the proper place in strings.xml

* Added confirm password field

when creating an account the create account password field must now also be filled in and match the password field. fields in create account now also display errors as you type.

* Refactored Database

username is now the unique ID and display name is now a field that will be able to be changed.

* Refactored LoginManager

organized methods in a logic manner and updated some method's logic.

* Added max length for usernames and passwords

With this update, the descriptions were updated and instead of using html, it was coded with SpannableStringBuilder

* Create account button added functionality

the create account button will now be disabled unless you have explicitly met all the criteria. The only way it will fail is if the username is taken or it can't connect to the server

---------

Co-authored-by: jhoyt15 <jacksonhoyt26@gmail.com>